aws/chalice_development/pipeline/requirements.txt
boto==2.49.0
chalice==1.27.3
numpy
scikit-learn==0.24.1
